+++ server/coingecko.mjs
@@ -1,4 +1,5 @@
 import fetch from "node-fetch"
+import {parseTime} from "./helpers.mjs";
 
 export const getPriceInfo = async (currency = 'usd') => {
     const url = `https://api.coingecko.com/api/v3/coins/markets?vs_currency=%CURRENCY%&ids=mina-protocol`.replace("%CURRENCY%", currency.toLowerCase())
@@ -17,6 +18,10 @@ export const getPriceInfo = async (currency = 'usd') => {
 
 export const processPriceInfo = async () => {
     const {currency, updateInterval} = globalThis.config.price
+    const _updateInterval = parseTime(updateInterval)
+
+    console.log("Get price interval", _updateInterval)
+
     let data = await getPriceInfo(currency)
 
     if (data) {
@@ -24,5 +29,5 @@ export const processPriceInfo = async () => {
         globalThis.priceInfo = data
     }
 
-    setTimeout(processPriceInfo, updateInterval)
+    setTimeout(processPriceInfo, _updateInterval)
 }