+++ server/helpers.mjs
@@ -3,6 +3,27 @@ import {discord} from "./discord.mjs"
 import {exec} from "child_process"
 import {hostname} from "os"
 
+export const parseTime = (t) => {
+    if (isNum(t)) return Math.abs(+t)
+    const pattern = /([0-9]+d)|([0-9]{1,2}h)|([0-9]{1,2}m)|([0-9]{1,2}s)/gm
+    const match = t.match(pattern)
+    return match.reduce( (acc, val) => {
+        let res
+
+        if (val.includes('d')) {
+            res = 1000 * 60 * 60 * 24 * parseInt(val)
+        } else if (val.includes('h')) {
+            res = 1000 * 60 * 60 * parseInt(val)
+        } else if (val.includes('m')) {
+            res = 1000 * 60 * parseInt(val)
+        } else if (val.includes('s')) {
+            res = 1000 * parseInt(val)
+        }
+
+        return acc + res
+    }, 0 )
+}
+
 export const parseTelegramChatIDs = s => s ? s.split(",").map( v => v.trim() ) : ""
 
 export const timestamp = () => {