+++ server/snark-worker-controller.mjs
@@ -1,5 +1,5 @@
 import {hostname} from "os"
-import {between, daemonStatus, execCommand, isNum, sendAlert} from "./helpers.mjs"
+import {between, daemonStatus, execCommand, isNum, parseTime, sendAlert} from "./helpers.mjs"
 
 export const processSnarkWorkerController = async () => {
     const config = globalThis.config.snarkWorker
@@ -11,13 +11,20 @@ export const processSnarkWorkerController = async () => {
     const {
         address,
         fee = 0.001,
-        stopBeforeBlock = 0,
-        startAfterBlock = 60000,
-        stopBeforeBlockTime = 300000,
+        stopBeforeBlock = "5m",
+        startAfterBlock = "1m",
         runWorkerCommand,
         setWorkerFeeCommand,
-        controlInterval = 10000
+        controlInterval = "10s"
     } = config
+
+    const _startAfterBlock = parseTime(startAfterBlock)
+    const _stopBeforeBlock = parseTime(stopBeforeBlock)
+    const _controlInterval = parseTime(controlInterval)
+
+    console.log("SW stop", _stopBeforeBlock)
+    console.log("SW start", _startAfterBlock)
+    console.log("SW control", _controlInterval)
 
     if (!address) return
 
@@ -28,7 +35,7 @@ export const processSnarkWorkerController = async () => {
         if (globalThis.snarkWorkerStopped === null || globalThis.snarkWorkerStopped) {
             let now = new Date().getTime()
 
-            if (!daemon.snarkWorker || (globalThis.snarkWorkerStoppedBlockTime && now > globalThis.snarkWorkerStoppedBlockTime + startAfterBlock)) {
+            if (!daemon.snarkWorker || (globalThis.snarkWorkerStoppedBlockTime && now > globalThis.snarkWorkerStoppedBlockTime + _startAfterBlock)) {
                 console.log("Start snark worker")
 
                 cmdFee = setWorkerFeeCommand.replace("<FEE>", fee)
@@ -50,10 +57,10 @@ export const processSnarkWorkerController = async () => {
         }
 
         if (nextBlock) {
-            if (stopBeforeBlock && !globalThis.snarkWorkerStopped) {
+            if (_stopBeforeBlock && !globalThis.snarkWorkerStopped) {
                 if (isNum(nextBlock) && nextBlock > 0) {
                     let now = new Date().getTime()
-                    let timeToStop = nextBlock - stopBeforeBlockTime
+                    let timeToStop = nextBlock - _stopBeforeBlock
                     let stop = between(now, timeToStop, nextBlock)
 
                     if (stop) {
@@ -73,5 +80,5 @@ export const processSnarkWorkerController = async () => {
         }
     }
 
-    setTimeout(processSnarkWorkerController, controlInterval)
+    setTimeout(processSnarkWorkerController, _controlInterval)
 }