+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 + [x] Server: added logic to disable `HANG` alert/restart when `hangInterval`, `hangIntervalAlert` is `0`
 + [x] Server: added control for mina stopped with `journalctl` and write this event to the log
 + [x] Server: added alert to `telegram` and/or `discord` about `mina daemon` stops by failure (info from `journalctl`)
++ [x] Server: added response for request `sync-state`. Return daemon sync state (`SYNCED`, `BOOTSTRAP`, ...) or `UNKNOWN`
++ [x] Server: added helper function `timeParser(t)` for parsing time, defined in short string (ex: `1d3h23m45s`, `30m`, `10s`, `2h30s`)
++ [x] Server: config parameter `restartAfterNotSynced` must be defined in milliseconds or short time string (ex: `30m`)
 
 ### 1.1.4
 + [x] Server: added `snark worker` controller